Bump lvmsync protocol version and update receiver
cc6ad8b changed the lvmsync protocol from expressing offsets in chunks
to bytes, but the protocol version and receiver code weren't updated,
resulting in incorrectly applied diffs. This patch updates the receiver
to correctly handle the new protocol, and also increases the protocol
version so as not to confuse older clients.