Parse the output of thin_dump line by line
For large volumes, the output of thin_dump can be tens of megabytes in
size, resulting in enormous memory usage when attempting to parse the
entire document as XML at once. This solution is less elegant, but it
gets the job done much more efficiently by regex matching one line at a
time.