Correct byte-order on LVM metadata reads [Fixes GH#31]
Somehow I managed to completely confuse myself as to the LVM metadata
format, and it was only the bug in htonq/ntohq (GH#6) that I fixed recently
that had made it work all this time.  Once I fixed that bug, this one
appeared.

Thanks to UnwashedMeme for the detective work, and pointing me in the right
direction.